+++ Realm/RLMSchema.mm
@@ -61,6 +61,23 @@ static NSMutableDictionary *s_localNameToClass = [[NSMutableDictionary alloc] in
         [schemas addObject:[cls sharedSchema]];
     }
     schema.objectSchema = schemas;
+
+    NSMutableArray *errors = [NSMutableArray new];
+    // Verify that all of the targets of links are included in the class list
+    for (RLMObjectSchema *objectSchema in schema->_objectSchema) {
+        for (RLMProperty *prop in objectSchema.properties) {
+            if (prop.type != RLMPropertyTypeObject && prop.type != RLMPropertyTypeArray) {
+                continue;
+            }
+            if (!schema->_objectSchemaByName[prop.objectClassName]) {
+                [errors addObject:[NSString stringWithFormat:@"- '%@.%@' links to class '%@', which is missing from the list of classes to persist", objectSchema.className, prop.name, prop.objectClassName]];
+            }
+        }
+    }
+    if (errors.count) {
+        @throw RLMException([@"Invalid class subset list:\n" stringByAppendingString:[errors componentsJoinedByString:@"\n"]]);
+    }
+
     return schema;
 }
 
@@ -90,14 +107,16 @@ static NSMutableDictionary *s_localNameToClass = [[NSMutableDictionary alloc] in
 }
 
 + (void)registerClasses:(Class *)classes count:(NSUInteger)count {
+    auto newClasses = [NSMutableArray new];
+    auto threadID = pthread_mach_thread_np(pthread_self());
+
     @synchronized(s_localNameToClass) {
-        auto threadID = pthread_mach_thread_np(pthread_self());
         // first create class to name mapping so we can do array validation
         // when creating object schema
         for (NSUInteger i = 0; i < count; i++) {
             Class cls = classes[i];
 
-            if (!RLMIsObjectSubclass(cls)) {
+            if (!RLMIsObjectSubclass(cls) || RLMIsGeneratedClass(cls)) {
                 continue;
             }
 
@@ -122,6 +141,7 @@ static NSMutableDictionary *s_localNameToClass = [[NSMutableDictionary alloc] in
             }
 
             s_localNameToClass[className] = cls;
+            [newClasses addObject:cls];
 
             RLMReplaceClassNameMethod(cls, className);
             // override sharedSchema class method to return nil to avoid topo-sort issues when on this thread
@@ -137,8 +157,8 @@ static NSMutableDictionary *s_localNameToClass = [[NSMutableDictionary alloc] in
             });
         }
 
-        NSMutableArray *schemas = [NSMutableArray arrayWithCapacity:s_localNameToClass.count];
-        for (Class cls in s_localNameToClass.allValues) {
+        NSMutableArray *schemas = [NSMutableArray arrayWithCapacity:newClasses.count];
+        for (Class cls in newClasses) {
             RLMObjectSchema *schema = [RLMObjectSchema schemaForObjectClass:cls];
 
             // set standalone class on shared shema for standalone object creation