+++ examples/installation/build.sh
@@ -30,34 +30,8 @@ EOF
 
 COMMAND="$1"
 
-prelaunch_simulator() {
-    killall "iOS Simulator" 2>/dev/null || true
-    killall Simulator 2>/dev/null || true
-    pkill CoreSimulator 2>/dev/null || true
-    # Erase all available simulators
-    (
-        IFS=$'\n' # make newlines the only separator
-        for LINE in $(xcrun simctl list); do
-            if [[ $LINE =~ unavailable ]]; then
-                # skip unavailable simulators
-                continue
-            fi
-            if [[ $LINE =~ ([0-9A-F]{8}-[0-9A-F]{4}-[0-9A-F]{4}-[0-9A-F]{4}-[0-9A-F]{12}) ]]; then
-                xcrun simctl erase "${BASH_REMATCH[1]}" 2>/dev/null || true
-            fi
-        done
-    )
-    sleep 5
-    if [[ -a "${DEVELOPER_DIR}/Applications/iOS Simulator.app" ]]; then
-        open "${DEVELOPER_DIR}/Applications/iOS Simulator.app"
-    elif [[ -a "${DEVELOPER_DIR}/Applications/Simulator.app" ]]; then
-        open "${DEVELOPER_DIR}/Applications/Simulator.app"
-    fi
-    sleep 5
-}
-
 xctest_ios() {
-    prelaunch_simulator
+    sh "$(dirname "$0")/../../scripts/reset-simulators.sh"
     XCODE_COMMAND="$@"
     xcodebuild $XCODE_COMMAND clean build test -sdk iphonesimulator || exit 1
 }